+++ docs/source/changes.md
@@ -6,6 +6,7 @@ Released April 2024
 
 ### New
 
+- [EHN] Add Harvard-Oxford atlas. (@htwangtw) [#117](https://github.com/bids-apps/giga_connectome/issues/117)
 - [DOCS] Improved documentation on using customised configuration files. (@htwangtw)
 - [ENH] use logger instead of print statements. (@Remi-Gau)
 
@@ -14,6 +15,8 @@ Released April 2024
 - [FIX] Bump nilearn version to 0.10.2 to fix issues [#26](https://github.com/bids-apps/giga_connectome/issues/26) and [#27](https://github.com/bids-apps/giga_connectome/issues/27). (@Remi-Gau)
 
 ### Enhancements
+
+- [ENH] Reduce the docker image size. (@htwangtw)
 
 ### Changes